requirements.txt
pymongo
sentence-transformers
neo4j
faiss-cpu
numpy
pandas
redis
sqlalchemy
pydantic